requirements.txt
# SmartRAG — core scientific stack
numpy>=1.24.0
scipy>=1.11.0
pandas>=2.0.0
scikit-learn>=1.3.0
tqdm>=4.66.0
pyyaml>=6.0
filelock>=3.13.0
requests>=2.31.0
regex>=2023.12.25

# Deep learning & Hugging Face ecosystem
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.40.0
datasets>=2.18.0
accelerate>=0.30.0
safetensors>=0.4.0
tokenizers>=0.15.0

# Project-specific retrieval, graphs, and RL
sentence-transformers>=3.0.0
spacy>=3.7.0
nltk>=3.8.0
rank_bm25>=0.2.2
faiss-cpu>=1.7.4
trl>=0.12.0
peft>=0.13.0
evaluate>=0.4.0
